+++ spring/src/main/java/cse416/spring/database/PrecinctWriter.java
@@ -43,7 +43,7 @@ public class PrecinctWriter {
     }
 
     // TODO Utilize PrecinctService to get all precincts
-    public static HashMap<Integer, Precinct> getAllPrecincts() {
+    public static HashMap<Long, Precinct> getAllPrecincts() {
         // Get all precincts
         EntityManagerFactory emf = Persistence.createEntityManagerFactory("orioles_db");
         EntityManager em = emf.createEntityManager();
@@ -51,7 +51,7 @@ public class PrecinctWriter {
         ArrayList<Precinct> allPrecincts = new ArrayList<>(precinctService.findAllPrecincts());
 
         // Convert the allPrecincts list into a hashmap of (id, precinct)
-        HashMap<Integer, Precinct> precinctHash = new HashMap<>();
+        HashMap<Long, Precinct> precinctHash = new HashMap<>();
 
         for (Precinct precinct : allPrecincts) {
             precinctHash.put(precinct.getId(), precinct);
